-reworked BlockStateData. -deleted ItemModelData and BlockModelData to rework then combine them into one class called ModelData.java -fixed a crash when creating new worlds. -Fuck you @Bear989Sr

+++ src/main/java/mightydanp/techcore/common/jsonconfig/trait/item/ItemTraitServer.java
@@ -68,47 +68,52 @@ public class ItemTraitServer extends JsonConfigServer<IItemTrait> {
     public Boolean isClientAndClientWorldConfigsSynced(Path singlePlayerConfigs){
         AtomicBoolean sync = new AtomicBoolean(true);
         Map<String, IItemTrait> clientItemTraits = new HashMap<>();
-
         Path path = Paths.get(singlePlayerConfigs + "/trait/item");
-        List<File> folders = Arrays.stream(Objects.requireNonNull(path.toFile().listFiles())).filter(file -> !file.getName().contains(".")).toList();
-        for(File folder : folders) {
 
-            File[] files = folder.listFiles();
+        File[] fileArray = path.toFile().listFiles();
 
-            if (files != null) {
-                if (getServerMap().size() != files.length) {
-                    sync.set(false);
-                    ConfigSync.syncedJson.put("item_trait", sync.get());
-                    return false;
-                }
+        if(fileArray != null){
+            List<File> folders = Arrays.stream(fileArray).filter(file -> !file.getName().contains(".")).toList();
 
-                if (files.length > 0) {
+            for (File folder : folders) {
 
-                    for (File file : files) {
-                        JsonObject jsonObject = TCJsonConfigs.itemTrait.getFirst().getJsonObject(file.getParentFile().getName() + ":" + file.getName().replace(".json", ""));
-                        IItemTrait itemTrait = ((ItemTraitRegistry) TCJsonConfigs.itemTrait.getFirst()).fromJsonObject(jsonObject);
-                        clientItemTraits.put(itemTrait.getRegistry(), itemTrait);
+                File[] files = folder.listFiles();
+
+                if (files != null) {
+                    if (getServerMap().size() != files.length) {
+                        sync.set(false);
+                        ConfigSync.syncedJson.put("item_trait", sync.get());
+                        return false;
                     }
 
-                    getServerMap().values().forEach(serverItemTrait -> {
-                        sync.set(clientItemTraits.containsKey(serverItemTrait.getRegistry()));
+                    if (files.length > 0) {
 
-                        if (sync.get()) {
-                            IItemTrait clientItemTrait = getServerMap().get(serverItemTrait.getRegistry());
-                            JsonObject jsonMaterial = ((ItemTraitRegistry) TCJsonConfigs.itemTrait.getFirst()).toJsonObject(serverItemTrait);
-                            JsonObject materialJson = ((ItemTraitRegistry) TCJsonConfigs.itemTrait.getFirst()).toJsonObject(clientItemTrait);
-
-                            sync.set(materialJson.equals(jsonMaterial));
+                        for (File file : files) {
+                            JsonObject jsonObject = TCJsonConfigs.itemTrait.getFirst().getJsonObject(file.getParentFile().getName() + ":" + file.getName().replace(".json", ""));
+                            IItemTrait itemTrait = ((ItemTraitRegistry) TCJsonConfigs.itemTrait.getFirst()).fromJsonObject(jsonObject);
+                            clientItemTraits.put(itemTrait.getRegistry(), itemTrait);
                         }
 
-                    });
-                }
+                        getServerMap().values().forEach(serverItemTrait -> {
+                            sync.set(clientItemTraits.containsKey(serverItemTrait.getRegistry()));
 
-            } else {
-                if (getServerMap().size() > 0) {
-                    sync.set(false);
-                    ConfigSync.syncedJson.put("item_trait", sync.get());
-                    return false;
+                            if (sync.get()) {
+                                IItemTrait clientItemTrait = getServerMap().get(serverItemTrait.getRegistry());
+                                JsonObject jsonMaterial = ((ItemTraitRegistry) TCJsonConfigs.itemTrait.getFirst()).toJsonObject(serverItemTrait);
+                                JsonObject materialJson = ((ItemTraitRegistry) TCJsonConfigs.itemTrait.getFirst()).toJsonObject(clientItemTrait);
+
+                                sync.set(materialJson.equals(jsonMaterial));
+                            }
+
+                        });
+                    }
+
+                } else {
+                    if (getServerMap().size() > 0) {
+                        sync.set(false);
+                        ConfigSync.syncedJson.put("item_trait", sync.get());
+                        return false;
+                    }
                 }
             }
         }
@@ -121,25 +126,29 @@ public class ItemTraitServer extends JsonConfigServer<IItemTrait> {
     public void syncClientWithServer(String folderName) throws IOException {
         Path path = Paths.get("config/" + Ref.mod_id + "/server" + "/trait/item");
 
-        List<File> folders = Arrays.stream(Objects.requireNonNull(path.toFile().listFiles())).filter(file -> !file.getName().contains(".")).toList();
+        File[] fileArray = path.toFile().listFiles();
 
-        for(File folder : folders) {
-            if (folder.listFiles() != null) {
-                for (File file : Objects.requireNonNull(folder.listFiles())) {
-                    Files.deleteIfExists(Paths.get(file.getAbsolutePath()));
+        if(fileArray != null) {
+            List<File> folders = Arrays.stream(fileArray).filter(file -> !file.getName().contains(".")).toList();
+
+            for (File folder : folders) {
+                if (folder.listFiles() != null) {
+                    for (File file : Objects.requireNonNull(folder.listFiles())) {
+                        Files.deleteIfExists(Paths.get(file.getAbsolutePath()));
+                    }
                 }
-            }
 
-            for (IItemTrait itemTrait : getServerMap().values()) {
-                String name = itemTrait.getRegistry();
-                Path file = Paths.get(folder + "/" + name + ".json");
-                JsonObject jsonObject = ((ItemTraitRegistry) TCJsonConfigs.itemTrait.getFirst()).toJsonObject(itemTrait);
-                String s = GSON.toJson(jsonObject);
-                if (!Files.exists(file)) {
-                    Files.createDirectories(file.getParent());
+                for (IItemTrait itemTrait : getServerMap().values()) {
+                    String name = itemTrait.getRegistry();
+                    Path file = Paths.get(folder + "/" + name + ".json");
+                    JsonObject jsonObject = ((ItemTraitRegistry) TCJsonConfigs.itemTrait.getFirst()).toJsonObject(itemTrait);
+                    String s = GSON.toJson(jsonObject);
+                    if (!Files.exists(file)) {
+                        Files.createDirectories(file.getParent());
 
-                    try (BufferedWriter bufferedwriter = Files.newBufferedWriter(file)) {
-                        bufferedwriter.write(s);
+                        try (BufferedWriter bufferedwriter = Files.newBufferedWriter(file)) {
+                            bufferedwriter.write(s);
+                        }
                     }
                 }
             }
@@ -150,9 +159,12 @@ public class ItemTraitServer extends JsonConfigServer<IItemTrait> {
         //Path serverConfigFolder = Paths.get("config/" + Ref.mod_id + "/server/" + folderName + "/material");
         Path path = Paths.get(folderName + "/trait/item");
         Path configFolder = Paths.get(TechAscension.mainJsonConfig.getFolderLocation()  + "/trait/item");
-        List<File> folders = Arrays.stream(Objects.requireNonNull(path.toFile().listFiles())).filter(file -> !file.getName().contains(".")).toList();
 
-        if (configFolder.toFile().listFiles() != null) {
+        File[] fileArray = path.toFile().listFiles();
+
+        if (configFolder.toFile().listFiles() != null && fileArray != null) {
+
+            List<File> folders = Arrays.stream(fileArray).filter(file -> !file.getName().contains(".")).toList();
             for(File folder : folders) {
                  if (folder.listFiles() == null){
                     for (File file : Objects.requireNonNull(configFolder.toFile().listFiles())) {
